fix(aislamiento): ampliar una guardia destapo dos IDOR vivos en el stock
Fases 1, 2, 8, 9 y 10 del hito 7 (TiendaNube), mas dos defectos que
aparecieron en el camino y no son del hito.

**Los dos IDOR.** Al ensanchar `analizarCreates` para que vea las formas
que no veia —la abreviada de ES6 y los routers que no se llaman
`router`— aparecio que POST /api/stock hacia
`Stock.findOrCreate({ where: { product_id, … } })` con el id del cuerpo
**sin que nadie mirara de quien es ese producto**. La fila salia con el
empresa_id de quien la mandaba —asi que revisando la tabla no se ve
nada raro— colgada del producto de otro cliente. Es la misma forma del
defecto de los pagos a proveedores: un create bajo un padre que nadie
valido.

Y POST /api/stock/bulk tiene el mismo defecto **y la guardia sigue sin
verlo**: reconoce `product_id: req.body.algo`, y ahi el id sale de un
elemento del arreglo, que es una indireccion mas. Por eso su test no es
redundante: es la unica red que existe para esa puerta. La validacion
va antes del bucle y sobre todos los ids juntos, porque una carga
masiva que escribe la mitad de las filas y falla en la otra deja un
inventario que nadie puede explicar.

**El doble de modelos no entendia `Op.in`.** `coincide` usa
Object.entries, que NO devuelve claves de tipo Symbol, y `Op.in` es un
Symbol: la consulta caia en la comparacion de igualdad, que compara un
numero contra un objeto y da false para toda fila. O sea que el doble
respondia «no encontre nada» no por los datos sino porque no entendia
la consulta, y cualquier codigo de produccion que use Op.in se
comportaba en los tests como si la tabla estuviera vacia. Es el modo de
falla que el encabezado de ese archivo advierte.

**Del hito:** src/controllers/ se disuelve y se borra —tenia un solo
archivo y NINGUNA de las tres guardias lo leia, asi que el IDOR de
createMapping nunca se miro aunque el detector reconoce exactamente esa
forma—; las dos migraciones con las cinco tablas nuevas; los cinco
modelos, con los ids de TiendaNube ensanchados a BIGINT porque int4
topa en 2.147 millones y esos identificadores los decide el tercero; y
las funciones puras de la web.

La semilla del verificador de reversibilidad gana settings,
puntos_de_venta, stock y tiendanube_mappings, mas la foto de datos de
esas tablas. Sin las dos mitades, la unica migracion con datos de este
hito **pasaba en verde sin ejecutar su rama**: el down no restauraba
ninguna fila y el script comparaba dos esquemas identicos. Comprobado
por mutacion.

`/api/tiendanube` sale de EXEMPT_PREFIXES y se parte en callback +
webhook, asi que las once rutas privadas quedan detras de
checkSubscription. El test que la spec pedia —«el webhook funciona con
la suscripcion vencida»— NO se escribio: pasaria igual con la lista
vacia, porque el webhook vive en el router publico y nunca llega a ese
middleware. Se escribio el contrario, que es el que puede fallar.

1647 tests de API + 79 de integracion, 874 en la web.

+++ apps/api/src/models/TiendanubeMapping.js
@@ -1,12 +1,31 @@
 const { DataTypes } = require('sequelize');
 const sequelize = require('../config/database');
 
+// ⚠ Los dos ids de TiendaNube son BIGINT desde la migración `20260810`, y eso es
+// una DESVIACIÓN DECLARADA del supuesto 4 de la spec 013 («el modelo y su
+// migración se conservan tal cual»). Lo que ese supuesto protege queda intacto:
+// los dos índices únicos, las dos FK con `ON DELETE CASCADE` y el bug documentado
+// de `addConstraint` con `key` en vez de `field`.
+//
+// El motivo: son identificadores de un tercero e `int4` topa en 2.147.483.647.
+// El síntoma del desbordamiento sería un 500 al insertar un mapeo, sin ningún
+// mensaje que diga qué pasó. Y dejarlos en INTEGER mientras
+// `tiendanube_variantes` los declara BIGINT sería peor: dos columnas del mismo
+// dato con anchos distintos, y el día del tope una tabla acepta la fila y la
+// otra no, dejando un mapeo que apunta a una variante que existe.
+//
+// `products.tiendanube_variant_id` **no** se ensancha: es la columna muerta que
+// deja de ser escribible (FR-070), y ensanchar una columna que nadie va a
+// escribir es trabajo sin destino.
+//
+// ⚠ El driver de Postgres devuelve `int8` como **string**: `mapping.tiendanube_variant_id === 123`
+// es `false` aunque el valor sea 123.
 const TiendanubeMapping = sequelize.define('TiendanubeMapping', {
   id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
   empresa_id: { type: DataTypes.INTEGER, allowNull: false },
   product_id: { type: DataTypes.INTEGER, allowNull: false },
-  tiendanube_variant_id: { type: DataTypes.INTEGER, allowNull: false },
-  tiendanube_product_id: { type: DataTypes.INTEGER, allowNull: false },
+  tiendanube_variant_id: { type: DataTypes.BIGINT, allowNull: false },
+  tiendanube_product_id: { type: DataTypes.BIGINT, allowNull: false },
 }, {
   tableName: 'tiendanube_mappings',
   timestamps: true,
